+++ lib/kitchen/directions/bake_exercises/v1.rb
@@ -21,7 +21,7 @@ module Kitchen::Directions::BakeExercises
             exercise_section: exercise_section, page: page, chapter: chapter)
 
           exercise_section.exercises.each do |exercise|
-            exercise.document.pantry(name: :link_text).store(
+            exercise.pantry(name: :link_text).store(
               "#{I18n.t(:exercise_label)} #{chapter.count_in(:book)}.#{exercise.count_in(:chapter)}",
               label: exercise.id
             )