+++ lib/kitchen/directions/move_exercises_to_eoc/v2.rb
@@ -22,7 +22,7 @@ module Kitchen::Directions::MoveExercisesToEOC
           # Get parent page title
           section_title = Kitchen::Directions::EocSectionTitleLinkSnippet.v1(page: page)
           exercise_section.exercises.each do |exercise|
-            exercise.document.pantry(name: :link_text).store(
+            exercise.pantry(name: :link_text).store(
               "#{I18n.t(:exercise_label)} #{chapter.count_in(:book)}.#{exercise.count_in(:chapter)}",
               label: exercise.id
             )