+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Access `.pantry` and `.clipboard` through an element instead of just its document (minor)
+
 ## [3.2.0] - 2021-04-19
 
 * Adds method to allow unit and page title text to be retrieved regardless of bake status (minor)